Move EFNYC-specific files into new efnyc dir.
