+++ src/serverless-functions/textbot.ts
@@ -1,4 +1,4 @@
-import { EfnycConversationHandlers } from "../textbot/efnyc-conversation-handlers";
+import { EfnycConversationHandlers } from "../textbot/efnyc/conversation-handlers";
 import { createServerlessHandlerForConversation } from "../textbot/run-as-serverless-function";
 import { serverlessRollbarHandler } from "../utils/serverless-util";
